feat: move fetchAvatars to fetch-avatar.ts
