+++ src/home/fetch-github/fetch-avatar.ts
@@ -3,6 +3,8 @@ import { getGitHubAccessToken } from "../getters/get-github-access-token";
 import { getImageFromCache, saveImageToCache } from "../getters/get-indexed-db";
 import { renderErrorInModal } from "../rendering/display-popup-modal";
 import { organizationImageCache } from "./fetch-issues-full";
+import { GitHubIssue } from "../github-types";
+import { taskManager } from "../home";
 
 // Map to track ongoing avatar fetches
 const pendingFetches: Map<string, Promise<Blob | void>> = new Map();
@@ -94,4 +96,19 @@ export async function fetchAvatar(orgName: string): Promise<Blob | void> {
     // Remove the pending fetch once it completes
     pendingFetches.delete(orgName);
   }
+}
+
+export async function fetchAvatars() {
+  const cachedTasks = taskManager.getTasks();
+
+  // fetches avatar for each organization for each task, but fetchAvatar() will only fetch once per organization, remaining are returned from cache
+  const avatarPromises = cachedTasks.map(async (task: GitHubIssue) => {
+    const [orgName] = task.repository_url.split("/").slice(-2);
+    if (orgName) {
+      return fetchAvatar(orgName);
+    }
+    return Promise.resolve();
+  });
+
+  await Promise.allSettled(avatarPromises);
 }
